+++ picocli-annotation-processing-tests/src/test/java/picocli/annotation/processing/tests/Issue1713Test.java
@@ -1,0 +1,26 @@
+package picocli.annotation.processing.tests;
+
+import com.google.testing.compile.Compilation;
+import com.google.testing.compile.JavaFileObjects;
+import org.junit.Test;
+
+import javax.annotation.processing.Processor;
+
+import static com.google.testing.compile.CompilationSubject.assertThat;
+import static com.google.testing.compile.Compiler.javac;
+
+public class Issue1713Test
+{
+    //@Ignore("https://github.com/remkop/picocli/issues/1713")
+    @Test
+    public void testIssue1713() {
+        Processor processor = new AnnotatedCommandSourceGeneratorProcessor();
+        Compilation compilation =
+                javac()
+                        .withProcessors(processor)
+                        .compile(JavaFileObjects.forResource(
+                                "picocli/issue1713/Command.java"));
+
+        assertThat(compilation).succeeded();
+    }
+}